+++ Presentation/RANSUROTTO.BLOG.Framework/Mvc/BaseModel.cs
@@ -46,7 +46,7 @@ namespace RANSUROTTO.BLOG.Framework.Mvc
     public partial class BaseEntityModel : BaseModel
     {
 
-        public virtual long Id { get; set; }
+        public virtual int Id { get; set; }
 
         public virtual Guid Guid { get; set; }
